+++ Classes/Extensions/Amazing+Tools.swift
@@ -20,7 +20,56 @@ public extension Amazing {
     }
     
     public func asAttributedText(fontSize: CGFloat, color: Color = Color.black, backgroundColor: Color = Color.clear) -> NSAttributedString {
-        return NSAttributedString.awesome(icon: self, fontSize: fontSize, color: color, backgroundColor: backgroundColor)
+        return NSAttributedString(icon: self, fontSize: fontSize, color: color, backgroundColor: backgroundColor)
+    }
+    
+}
+
+extension Amazing {
+    
+    public var key: String {
+        return String(reflecting: self)
+            .split(separator: ".")
+            .suffix(3)
+            .map { String($0).camelCaseToString(with: "-").lowercased() }
+            .joined(separator: ".")
+    }
+    
+    public var description: String {
+        return String(describing: self).camelCaseToString()
+    }
+    
+    public static var allKeys: [String] {
+        return allCases.map { $0.key }
+    }
+    
+    public static var allDescriptions: [String] {
+        return allCases.map { $0.description }
+    }
+    
+    @available(swift, deprecated: 4.2, message: "This will be removed in the future. Please use .allKeys instead")
+    public static var keys: [String] {
+        return allKeys
+    }
+    
+    @available(swift, deprecated: 4.2, message: "This will be removed in the future. Please use .allDescriptions instead")
+    public static var labels: [String] {
+        return allDescriptions
+    }
+    
+    @available(swift, deprecated: 4.2, message: "This will be removed in the future. Please use .rawValue instead.")
+    public var code: String {
+        return self.rawValue
+    }
+    
+    @available(swift, deprecated: 4.2, message: "This will be removed in the future. Please use .key instead.")
+    public var name: String {
+        return String(describing: self).camelCaseToString(with: "-").lowercased()
+    }
+    
+    @available(swift, deprecated: 4.2, message: "This will be removed in the future. Please use .description instead.")
+    public var label: String {
+        return self.description
     }
     
 }
